chore: types for renderer moved to types file
